+++ internal/integration_test/bench/decoder_bench_test.go
@@ -1,0 +1,20 @@
+package bench
+
+import (
+	"testing"
+
+	"github.com/tetratelabs/wazero/api"
+	"github.com/tetratelabs/wazero/internal/wasm"
+	"github.com/tetratelabs/wazero/internal/wasm/binary"
+)
+
+func BenchmarkCodec(b *testing.B) {
+	b.Run("binary.DecodeModule", func(b *testing.B) {
+		b.ReportAllocs()
+		for i := 0; i < b.N; i++ {
+			if _, err := binary.DecodeModule(caseWasm, api.CoreFeaturesV2, wasm.MemoryLimitPages, false, false); err != nil {
+				b.Fatal(err)
+			}
+		}
+	})
+}
